adds repositories and updates db:test. Also updates migration to eliminate BIGSERIAL which can't be represented by ts integers.

+++ db/migrations/001_initial_schema.sql
@@ -2,31 +2,31 @@
 -- Tables: users, resumes, job_descriptions, analysis_runs
 
 CREATE TABLE users (
-  id BIGSERIAL PRIMARY KEY,
+  id SERIAL PRIMARY KEY,
   username TEXT NOT NULL UNIQUE,
   created_at TIMESTAMPTZ NOT NULL DEFAULT NOW()
 );
 
 CREATE TABLE resumes (
-  id BIGSERIAL PRIMARY KEY,
-  user_id BIGINT NOT NULL REFERENCES users(id) ON DELETE CASCADE,
+  id SERIAL PRIMARY KEY,
+  user_id INTEGER NOT NULL REFERENCES users(id) ON DELETE CASCADE,
   resume_text TEXT NOT NULL DEFAULT '',
   created_at TIMESTAMPTZ NOT NULL DEFAULT NOW(),
   updated_at TIMESTAMPTZ NOT NULL DEFAULT NOW()
 );
 
 CREATE TABLE job_descriptions (
-  id BIGSERIAL PRIMARY KEY,
-  user_id BIGINT NOT NULL REFERENCES users(id) ON DELETE CASCADE,
+  id SERIAL PRIMARY KEY,
+  user_id INTEGER NOT NULL REFERENCES users(id) ON DELETE CASCADE,
   job_text TEXT NOT NULL DEFAULT '',
   created_at TIMESTAMPTZ NOT NULL DEFAULT NOW(),
   updated_at TIMESTAMPTZ NOT NULL DEFAULT NOW()
 );
 
 CREATE TABLE analysis_runs (
-  id BIGSERIAL PRIMARY KEY,
-  resume_id BIGINT NOT NULL REFERENCES resumes(id) ON DELETE CASCADE,
-  job_description_id BIGINT NOT NULL REFERENCES job_descriptions(id) ON DELETE CASCADE,
+  id SERIAL PRIMARY KEY,
+  resume_id INTEGER NOT NULL REFERENCES resumes(id) ON DELETE CASCADE,
+  job_description_id INTEGER NOT NULL REFERENCES job_descriptions(id) ON DELETE CASCADE,
   ai_result_json JSONB NOT NULL,
   model_metadata JSONB,
   created_at TIMESTAMPTZ NOT NULL DEFAULT NOW()
